chore: use `throws` and `rejects` assertions (#36)
<!-- What / Why -->
<!-- Describe the request in detail. What it does and why it's being
changed. -->

When testing that a function throws an error, it's easy to forget to
include the `fail` assertion in the try block, and end up with a test
that verifies nothing. I find `throws` and `rejects` assertions to be
less error prone, and try to use them instead of assertions in a catch
block.

Caveat: Unfortunately with tap 16, you can supply an object for
comparison with the thrown error, but can't run arbitrary assertions.
`throws` and `rejects` return the thrown error in newer versions of tap.

## References
<!-- Examples:
  Related to #0
  Depends on #0
  Blocked by #0
  Fixes #0
  Closes #0
-->